refactor vec2/math
