+++ common/common.vcxitems
@@ -30,7 +30,7 @@
     <ClInclude Include="$(MSBuildThisFileDirectory)rawaccel-version.h" />
     <ClInclude Include="$(MSBuildThisFileDirectory)rawaccel.hpp" />
     <ClInclude Include="$(MSBuildThisFileDirectory)utility-install.hpp" />
-    <ClInclude Include="$(MSBuildThisFileDirectory)vec2.h" />
+    <ClInclude Include="$(MSBuildThisFileDirectory)math-vec2.hpp" />
     <ClInclude Include="$(MSBuildThisFileDirectory)utility.hpp" />
   </ItemGroup>
 </Project>
