--- common/vec2.h
@@ -1,9 +1,0 @@
-#pragma once
-
-template <typename T>
-struct vec2 {
-    T x;
-    T y;
-};
-
-using vec2d = vec2<double>;
